requirements.txt
# Discord bot framework
py-cord>=2.6.0
audioop-lts>=0.2.0  # Required for py-cord on Python 3.13+

# Gmail API integration
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Configuration management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0

# Linting
ruff>=0.6.0
